+++ archetype-pom/pom-project/src/main/resources/META-INF/archetype-post-generate.groovy
@@ -36,6 +36,10 @@ def removeRootProjectLaunchers() {
 def removeRootProjectPomConfiguration() {
     projectPom = projectDir.resolve("pom.xml")
     out = Files.readAllLines(projectPom)
+    // Remove SCM configuration
+    for (int i = 60; i >= 56; i--) {
+        out.remove(i);
+    }
     // Remove maven-site-plugin configuration
     for (int i = 55; i >= 42; i--) {
         out.remove(i);
@@ -43,6 +47,8 @@ def removeRootProjectPomConfiguration() {
     // Remove Sonar project key configuration
     out.remove(37);
     out.remove(36);
+    // Remove project URL configuration
+    out.remove(20);
     Files.write(projectPom, out, StandardOpenOption.WRITE, StandardOpenOption.TRUNCATE_EXISTING)
 }